+++ test/entities/dummy.dart
@@ -1,4 +1,0 @@
-class Dummy {
-  @override
-  String toString() => 'Dummy';
-}